match เร็วขึ้น ~95% :D

```bash
$ benchcmp before.out after.out
benchmark                  old ns/op     new ns/op     delta
Benchmark_Seg_isThai-8     7648          366           -95.21%
```